move trpc server files to server only dir
