+++ package.json
@@ -1,35 +1,36 @@
 {
-	"name": "polls",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"dev": "vite dev",
-		"build": "vite build",
-		"preview": "vite preview",
-		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
-	},
-	"devDependencies": {
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/kit": "^1.15.7",
-		"@types/bcrypt": "^5.0.0",
-		"autoprefixer": "^10.4.14",
-		"postcss": "^8.4.22",
-		"prisma": "^4.13.0",
-		"svelte": "^3.58.0",
-		"svelte-check": "^3.2.0",
-		"tailwindcss": "^3.3.1",
-		"tslib": "^2.5.0",
-		"typescript": "^5.0.4",
-		"vite": "^4.2.2"
-	},
-	"type": "module",
-	"dependencies": {
-		"@prisma/client": "^4.13.0",
-		"@trpc/client": "^10.20.0",
-		"@trpc/server": "^10.20.0",
-		"bcrypt": "^5.1.0",
-		"trpc-sveltekit": "^3.5.5",
-		"zod": "^3.21.4"
-	}
+  "name": "polls",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "dev": "vite dev",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+  },
+  "devDependencies": {
+    "@sveltejs/adapter-auto": "^2.0.0",
+    "@sveltejs/kit": "^1.15.7",
+    "@types/bcrypt": "^5.0.0",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.22",
+    "prisma": "^4.13.0",
+    "svelte": "^3.58.0",
+    "svelte-check": "^3.2.0",
+    "tailwindcss": "^3.3.1",
+    "tslib": "^2.5.0",
+    "typescript": "^5.0.4",
+    "vite": "^4.2.2"
+  },
+  "type": "module",
+  "dependencies": {
+    "@prisma/client": "^4.13.0",
+    "@sveltejs/adapter-node": "^1.2.3",
+    "@trpc/client": "^10.20.0",
+    "@trpc/server": "^10.20.0",
+    "bcrypt": "^5.1.0",
+    "trpc-sveltekit": "^3.5.5",
+    "zod": "^3.21.4"
+  }
 }
