+++ src/hooks.server.ts
@@ -1,5 +1,5 @@
-import { createContext } from "$lib/trpc/context";
-import { router } from "$lib/trpc/router";
+import { createContext } from "$lib/server/trpc/context";
+import { router } from "$lib/server/trpc/router";
 import type { Handle } from "@sveltejs/kit";
 import { createTRPCHandle } from "trpc-sveltekit";
