+++ src/lib/server/trpc/t.ts
@@ -1,4 +1,4 @@
-import type { Context } from "$lib/trpc/context";
+import type { Context } from "$lib/server/trpc/context";
 import { initTRPC } from "@trpc/server";
 
 export const t = initTRPC.context<Context>().create();
