+++ src/lib/trpc-client/client.ts
@@ -1,4 +1,4 @@
-import type { Router } from "$lib/trpc/router";
+import type { Router } from "$lib/server/trpc/router";
 import { createTRPCClient, type TRPCClientInit } from "trpc-sveltekit";
 
 let browserClient: ReturnType<typeof createTRPCClient<Router>>;
