+++ src/routes/account/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { page } from "$app/stores";
-	import { tAuthSafe } from "$lib/trpc/autoRedirect";
+	import { tAuthSafe } from "$lib/trpc-client/autoRedirect";
 	import type { PageData } from "./$types";
 
 	export let data: PageData;
