+++ src/routes/account/+page.ts
@@ -1,4 +1,4 @@
-import { tAuthSafe } from "$lib/trpc/autoRedirect";
+import { tAuthSafe } from "$lib/trpc-client/autoRedirect";
 import type { PageLoad } from "./$types";
 
 export const load: PageLoad = async (event) =>
